cloud_storage: move configuration methods to own TU
Prepare for splitting types lib.
